+++ src/trio_util/_async_value.py
@@ -1,4 +1,6 @@
-from collections import defaultdict
+from collections import defaultdict, namedtuple
+from contextlib import asynccontextmanager
+from functools import partial
 
 import trio
 
@@ -163,3 +165,48 @@ class AsyncValue:
         returns (value, old_value) which satisfied the predicate
         """
         return await self._wait_predicate(self._edge_results, _ValueWrapper(value_or_predicate))
+
+
+@asynccontextmanager
+async def compose_values(**value_map):
+    """Async context manager providing a composite of multiple AsyncValues
+
+    The composite object itself is an AsyncValue, with the `value` of each
+    underlying object accessible as attributes on the composite `value`.
+
+    `compose_values()` expects named AsyncValue instances to be provided as
+    keyword arguments.  The attributes of the composite value will correspond
+    to the given names.
+
+    It's mostly an implementation detail, but the composite value type is a
+    namedtuple.  Users should not write to the composite `value` attribute
+    since it is exclusively managed by the context.
+
+    Synopsis:
+        >>> async_x, async_y = AsyncValue(-1), AsyncValue(10)
+        >>>
+        >>> async with compose_values(x=async_x, y=async_y) as async_xy:
+        >>>     result = await async_xy.wait_value(lambda val: val.x < 0 < val.y))
+        >>>
+        >>> result
+        CompositeValue(x=-1, y=10)
+    """
+    async with trio.open_nursery() as nursery:
+        values = value_map.values()
+        # the implementation requires `wait_transition()` taking a predicate
+        if not (values and all(isinstance(e, AsyncValue) for e in values)):
+            raise TypeError('expected instances of AsyncValue')
+        value_type = namedtuple('CompositeValue', value_map.keys())
+        composite = AsyncValue(value_type._make(e.value for e in values))
+
+        # This predicate is run indefinitely per underlying value and updates
+        # the composite as a side effect.
+        def _update_composite(name, val, _):
+            composite.value = composite.value._replace(**{name: val})
+            return False
+
+        for name, e in value_map.items():
+            nursery.start_soon(e.wait_transition, partial(_update_composite, name))
+
+        yield composite
+        nursery.cancel_scope.cancel()